refactor: svgDocument를 조작하는 클래스들을 분리한다. (#37)
* refactor: Count class 분리

* refactor: ElementManipulator 역할에 맞게 클래스 분리

* refactor: 변수명 수정

* refactor: emoji viewBox 수정하는 부분으로 변경

* fix: db에 저장할때 ""부분 제거하도록 변경

+++ click-service/src/main/java/clickme/clickme/repository/HeartRedisRepository.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Repository;
 
 import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 @Repository
 @Conditional(RedisConnectionCondition.class)
@@ -45,7 +46,7 @@ public class HeartRedisRepository implements HeartRepository {
 
     @Override
     public List<RankingResponse> findRealTimeRanking(final int start, final int end) {
-        final AtomicInteger ranking = new AtomicInteger(start);
+        final AtomicLong ranking = new AtomicLong(start);
         return zSet.reverseRangeWithScores(KEY, start, end)
                 .stream()
                 .map(tuple -> new RankingResponse(ranking.getAndIncrement(), tuple.getValue(), tuple.getScore().longValue()))
